requirements.txt
requests==2.31.0
websocket-client==1.6.3
pandas==2.1.1
numpy==1.24.3
TA-Lib==0.4.25
asyncio==3.4.3
python-dateutil==2.8.2
sqlite3 